requirements.txt
Django==4.2
djangorestframework==3.14
psycopg2-binary==2.9.9
python-keycloak==3.7
PyJWT==2.8.0